Move lit-html.ts functionality into core
