+++ src/test/lib/lit-extended_test.ts
@@ -15,7 +15,7 @@
 /// <reference path="../../../node_modules/@types/mocha/index.d.ts" />
 /// <reference path="../../../node_modules/@types/chai/index.d.ts" />
 
-import {AttributePart, directive, html as htmlPlain} from '../../core.js';
+import {AttributePart, directive, html as htmlPlain} from '../../index.js';
 import {html, render} from '../../lib/lit-extended.js';
 import {stripExpressionDelimeters} from '../test-helpers.js';
